requirements.txt

# FastAPI y servidor
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Base de datos (sin versiones específicas para evitar compilación)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Excel y Data
pandas>=2.1.0
openpyxl>=3.1.0

# PDFs
pymupdf>=1.23.0
pillow>=10.0.0

# Emails
sib-api-v3-sdk>=7.6.0

# Google Drive
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# Sincronización automática
apscheduler>=3.10.0

# Utilidades
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Para descargar PDFs desde Drive (NUEVO)
requests>=2.31.0

# PDF editing y mejora de calidad (AGREGAR AL FINAL)
pdf2image>=1.16.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
reportlab>=4.0.0
# ✅ NUEVO: Anthropic para Claude Haiku
anthropic==0.40.0